Fix build warnings
For src/gui/ui_device.c, uint32_t is likely the correct type to use and Moonlight probably should fix their header. I'm just too lazy to create a PR for now.

+++ src/connection.c
@@ -136,7 +136,7 @@ void connection_stage_complate(int stage) {
   vita_debug_log("connection_stage_complate - stage: %d\n", stage);
 }
 
-void connection_stage_failed(int stage, long code) {
+void connection_stage_failed(int stage, int code) {
   connection_failed_stage = stage;
   connection_failed_stage_code = code;
   vita_debug_log("connection_stage_failed - stage: %d, %d\n", stage, code);
